requirements.txt
asgiref==3.4.0
Django==3.2.4
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-mptt==0.12.0
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
requests~=2.25.1